Add credentials header into fetch request
The default value for credentials is "same-origin", however this wasn't always the case. Older browsers implemented a version of the fetch specification where the default was "omit".  Because of this, older browsers omit the cookie when making requests, causing a 401 to be returned from the server

See https://github.com/github/fetch#receiving-cookies